requirements.txt
azure-common==1.1.25
azure-core==1.5.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
azure-batch==9.0.0
azure-storage-blob==12.3.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
cryptography==2.9.2
cycler==0.10.0
descartes==1.1.0
filelock==3.0.12
idna==2.9
isodate==0.6.0
kiwisolver==1.2.0
matplotlib==3.1.3
numpy==1.18.1
oauthlib==3.1.0
palettable==3.3.0
pandas==1.0.3
patsy==0.5.1
psutil==5.7.0
pycparser==2.20
pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
scipy==1.4.1
six==1.14.0
statsmodels==0.11.1
tornado==6.0.4
urllib3==1.25.9
